borrar version por si acaso

+++ controllers/ControllerBase.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace PHPMaker2022\project11;
+namespace PHPMaker2023\project11;
 
 use Psr\Container\ContainerInterface;
 use Psr\Http\Message\ServerRequestInterface as Request;
@@ -21,7 +21,7 @@ class ControllerBase
     }
 
     // Run page
-    protected function runPage(Request $request, Response $response, array $args, string $pageName, bool $useLayout = true): Response
+    protected function runPage(Request $request, Response $response, array $args, string $pageName, string $viewName = null, bool $useLayout = true): Response
     {
         global $RouteValues;
 
@@ -55,7 +55,7 @@ class ControllerBase
                 if (
                     !$page->UseLayout || // No layout
                     property_exists($page, "IsModal") && $page->IsModal || // Modal
-                    $request->getQueryParam(Config("PAGE_LAYOUT")) !== null // Multi-Column List page
+                    $request->getParam(Config("PAGE_LAYOUT")) !== null // Multi-Column List page
                 ) { // Partial view
                     $useLayout = false;
                 }
@@ -65,12 +65,75 @@ class ControllerBase
 
                 // Render view with $GLOBALS
                 $page->RenderingView = true;
-                $template = $page->View ?? $pageName . ".php"; // View
-                $GLOBALS["Title"] = $GLOBALS["Title"] ?? $page->Title; // Title
+                $template = ($page->View ?? $viewName ?? $pageName) . ".php"; // View
+                $GLOBALS["Title"] ??= $page->Title; // Title
                 try {
                     $response = $view->render($response, $template, $GLOBALS);
                 } finally {
                     $page->RenderingView = false;
+                    $page->terminate(); // Terminate page and clean up
+                }
+            }
+
+            // Clean up temp folder if not add/edit/export
+            if (
+                property_exists($page, "TableName") && // Table/Report class
+                !in_array($page->PageID, ["add", "register", "edit", "update"]) && // Not add/register/edit/update page
+                !($page->PageID == "list" && $page->isAddOrEdit()) && // Not list page add/edit
+                !(property_exists($page, "Export") && $page->Export != "" && $page->Export != "print" && $page->UseCustomTemplate) // Not export custom template
+            ) {
+                CleanUploadTempPaths(session_id());
+            }
+            return $response;
+        }
+
+        // Page not found
+        throw new HttpNotFoundException($request);
+    }
+
+    // Run chart
+    protected function runChart(Request $request, Response $response, array $args, string $pageName, string $chartVar): Response
+    {
+        global $RouteValues;
+
+        // Route values
+        // Note:
+        // - $RouteValues[0] set up in PermissionMiddleWare
+        // - $chartVar in Route(1)
+        $RouteValues = array_merge($RouteValues, [$chartVar], $args, array_values($args));
+
+        // Generate new CSRF token
+        GenerateCsrf();
+
+        // Create page
+        $pageClass = PROJECT_NAMESPACE . $pageName;
+        if (class_exists($pageClass)) {
+            // Set up response object
+            $GLOBALS["Response"] = &$response; // Note: global $Response does not work
+
+            // Create page object
+            $page = new $pageClass();
+            $GLOBALS["Page"] = &$page;
+
+            // Write header
+            $cache = Config("CACHE"); // No cache for preview
+            WriteHeader($cache);
+
+            // Run the page
+            $page->run();
+
+            // Render chart
+            if (property_exists($page, $chartVar)) {
+                $chart = $page->$chartVar;
+
+                // Output chart
+                try {
+                    $chartClass = ($chart->PageBreakType == "before") ? "ew-chart-bottom" : "ew-chart-top";
+                    $chartWidth = $request->getQueryParam("width");
+                    $chartHeight = $request->getQueryParam("height");
+                    $html = $chart->render($chartClass, $chartWidth, $chartHeight);
+                    Write($html);
+                } finally {
                     $page->terminate(); // Terminate page and clean up
                 }
             }
